Refactor Merkle tree functions
